Move FetchJetpackStatus to better location
